requirements.txt
pyannote.audio==3.2.0
pydub==0.25.1
typer==0.12.3
feedparser==6.0.11
requests==2.31.0
